+++ src/react/components/app/grid-view/styles.css
