+++ src/react/components/properties-filter-app/group-tag-list/styles.css
