+++ src/react/components/properties-filter-app/group-tag/index.tsx
