+++ src/react/components/properties-filter-app/group-tag/styles.css
