+++ src/react/components/shared/divider/index.tsx
