+++ src/react/components/shared/property/styles.css
