+++ src/react/components/shared/stack/index.tsx
