+++ src/react/components/shared/switch/index.tsx
