+++ src/react/components/shared/switch/styles.css
