+++ src/react/components/shared/tag/styles.css
